SDK-334 move sdkfactory and stopwatch to own file
